Removed var_states() function.
The function was previously defined inside the get_desired_data() function and was used for testing and/or dealing with variable states for checkboxes.
After testing, the function did nothing besides master.destroy() which was effectively completed with the "Cancel" button.  So the function was removed and the cancel button was renamed "Show graphs".